Move CustomShortcutModel definition to a new file
1. Move CustomShortcutModel definition to a new file
2. Some other minor tweaks
